Asymptotic Covariance Matrix for Ordinary Least Squares Regression

Calculates the asymptotic covariance matrix of the unique elements of the covariance matrix using ordinary least squares estimates.

gammacap_ols(
  x,
  type = "gen",
  adf_unbiased = TRUE,
  bcap = 5000L,
  seed = NULL,
  ke_unbiased = FALSE,
  yc = FALSE,
  names = TRUE,
  sep = "."
)

Arguments

x
Object of class lm.

type
Character string. Type of asymptotic covariance matrix of the regressors If type = "adf", calculate asymptotic distribution-free covariance matrix. If type = "adfnb", calculate nonparametric bootstrapped asymptotic distribution-free covariance matrix. If type = "gen", calculate covariance matrix using the general formula. If type = "mvn", calculate covariance matrix with multivariate normal data.

adf_unbiased
Logical. If adf_unbiased = TRUE, use unbiased asymptotic distribution-free covariance matrix. If adf_unbiased = FALSE, use consistent asymptotic distribution-free covariance matrix.

bcap
Integer. Number of bootstrap samples when type = "gen" and type = "adfnb".

seed
Integer. Random number generation seed when type = "gen" and type = "adfnb".

ke_unbiased
Logical. If ke_unbiased = TRUE, use estimator \( \frac{ \frac{1}{n} \sum_{i=1}^{n} \hat{\varepsilon}_{i}^{4} }{ \left( \frac{1}{n - p - 1} \sum_{i=1}^{n} \hat{\varepsilon}_{i}^{2} \right)^2 } \) where the denominator is the square of the unbiased estimator of the error variance. If ke_unbiased = FALSE, use estimator \( \frac{ \frac{1}{n} \sum_{i=1}^{n} \hat{\varepsilon}_{i}^{4} }{ \left( \frac{1}{n} \sum_{i=1}^{n} \hat{\varepsilon}_{i}^{2} \right)^2 } \) where the denominator is the square of the consistent estimator of the error variance.

yc
Logical. If yc = TRUE, order the output following Yuan and Chan (2011) page 674. If yc = FALSE, the order of the output following \(\mathrm{vech} \left( \Sigma \right)\).

names
Logical. Add names.

sep
Character string. Separator for variable names.

Value

A matrix.

References

Yuan, K.-H., & Chan, W. (2011). Biases and standard errors of standardized regression coefficients. Psychometrika, 76, 670–690. doi:10.1007/S11336-011-9224-6.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

set.seed(42)
n <- 1000
k <- 2
z <- matrix(
  data = rnorm(n = n * k), nrow = n, ncol = k
)
q <- chol(
  matrix(
    data = c(1.0, 0.5, 0.5, 1.0),
    nrow = k, ncol = k
  )
)
x <- as.data.frame(z %*% q)
colnames(x) <- c("y", "x")
obj <- lm(y ~ x, data = x)

gammacap_ols(obj)

#>           y.y      y.x       x.x
#> y.y 2.0240018 1.049240 0.5633753
#> y.x 1.0492404 1.296956 1.0902684
#> x.x 0.5633753 1.090268 2.1099351