Hessian Matrix of the Multivariate Normal Distribution - Generic

Calculates hessian matrix of the log of the likelihood function of the multivariate normal distribution for the ith observation.

hess_l_mvn_generic(x, mu, sigmacap)

Arguments

x
Numeric vector of length k. The ith vector of observations.

mu
Numeric vector. Parameter. Mean vector \(\boldsymbol{\mu}\).

sigmacap
Numeric matrix. Parameter. Covariance matrix \(\boldsymbol{\Sigma}\).

Value

A matrix.

See also

Examples

n <- 5
mu <- c(0, 0)
sigmacap <- matrix(
  data = c(
    1, 0.5, 0.5, 1
  ),
  nrow = 2
)

xcap <- as.data.frame(
  t(
    rmvn_chol(
      n = n,
      mu = mu,
      sigmacap = sigmacap
    )
  )
)

lapply(
  X = xcap,
  FUN = hess_l_mvn_generic,
  mu = mu,
  sigmacap = sigmacap
)

#> $V1
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] -1.3333333  0.6666667  0.4178355 -0.8011668  0.2961245
#> [2,]  0.6666667 -1.3333333 -0.2089178  0.7139601 -0.5922490
#> [3,]  0.4178355 -0.2089178  0.7579490 -0.6378219  0.1294237
#> [4,] -0.8011668  0.7139601 -0.6378219  1.6426161 -0.5717573
#> [5,]  0.2961245 -0.5922490  0.1294237 -0.5717573  0.6258197
#> 
#> $V2
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] -1.3333333  0.6666667 -0.3776310  2.0450453 -0.9281149
#> [2,]  0.6666667 -1.3333333  0.1888155 -1.3057459  1.8562298
#> [3,] -0.3776310  0.1888155  0.7819350 -0.3096845 -0.0406415
#> [4,]  2.0450453 -1.3057459 -0.3096845 -0.9946510  0.9289345
#> [5,] -0.9281149  1.8562298 -0.0406415  0.9289345 -1.6953030
#> 
#> $V3
#>            [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] -1.3333333  0.6666667  1.9635212 -2.825973  0.9221063
#> [2,]  0.6666667 -1.3333333 -0.9817606  2.885628 -1.8442126
#> [3,]  1.9635212 -0.9817606 -2.0026728  3.272755 -1.1357092
#> [4,] -2.8259732  2.8856275  3.2727549 -5.936043  3.1023941
#> [5,]  0.9221063 -1.8442126 -1.1357092  3.102394 -1.6619512
#> 
#> $V4
#>            [,1]       [,2]       [,3]      [,4]       [,5]
#> [1,] -1.3333333  0.6666667  1.5140447 -2.744864  0.9939210
#> [2,]  0.6666667 -1.3333333 -0.7570224  2.507966 -1.9878420
#> [3,]  1.5140447 -0.7570224 -0.8303597  2.227997 -0.9064084
#> [4,] -2.7448644  2.5079657  2.2279967 -4.717925  2.8501909
#> [5,]  0.9939210 -1.9878420 -0.9064084  2.850191 -2.0747480
#> 
#> $V5
#>            [,1]       [,2]        [,3]      [,4]        [,5]
#> [1,] -1.3333333  0.6666667  0.32321889 -2.444948  1.14166905
#> [2,]  0.6666667 -1.3333333 -0.16160944  1.464888 -2.28333810
#> [3,]  0.3232189 -0.1616094  0.81053605 -0.296199 -0.05453453
#> [4,] -2.4449475  1.4648879 -0.29619897 -2.319869  1.61973695
#> [5,]  1.1416691 -2.2833381 -0.05453453  1.619737 -3.02133578
#>